requirements.txt
absl-py==0.8.0
astor==0.8.0
backcall==0.1.0
brewer2mpl==1.4.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cloudpickle==1.2.1
cycler==0.10.0
decorator==4.4.0
disentanglement-lib==1.2
future==0.17.1
gast==0.2.2
gin-config==0.2.1
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.9.0
idna==2.8
imageio==2.5.0
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
joblib==0.13.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
numpy==1.17.1
olefile==0.46
pandas==0.25.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==2.0.9
protobuf==3.9.1
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.4.2
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
pyzmq==18.1.0
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.1
simplejson==3.16.0
six==1.12.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-hub==0.6.0
tensorflow-probability==0.7.0
termcolor==1.1.0
torch==1.2.0
torchfile==0.1.0
torchvision==0.4.0a0+6b959ee
tornado==6.0.3
tqdm==4.35.0
traitlets==4.3.2
urllib3==1.25.3
visdom==0.1.8.8
wcwidth==0.1.7
websocket-client==0.56.0
Werkzeug==0.15.6
wrapt==1.11.2
